Combined new assertions and old
